Cast some odd types to uint
For ABOUT functions used in bugworkaround.cpp and devcons.cpp

+++ Main/Source/bugworkaround.cpp
@@ -277,8 +277,8 @@ bool bugfixdp::ScanLevelForCharactersAndItemsWork(
       ABORT( //more than 2 copies (one valid and one invalid) is more complicated and the code is not ready yet to deal with that.
         "iDupCount=%d found for item 0x%x id=%lu."
         "tot=%d item pointer/object/reference 0x%x found for id=%lu.",
-        iDupIDCount,itWork,itWork->GetID(),
-        iPointerMatchCount,itWork,itWork->GetID()
+        iDupIDCount,(uint)itWork,itWork->GetID(),
+        iPointerMatchCount,(uint)itWork,itWork->GetID()
       );
 
   return iDupIDCount>0;
